third_party/blink/web_tests/external/wpt/content-security-policy/inside-worker/dedicatedworker-report-only-expected.txt

This is a testharness.js-based test.
[FAIL] WebSocket.
  promise_test: Unhandled rejection with value: "timeout while waiting for report for https://web-platform.test:9444/echo"
[FAIL] connect-src-self-report-only
  assert_array_equals: Reports do not match expected property 1 to be "https://web-platform.test:9444/echo" but got "https://www.web-platform.test:8445/content-security-policy/support/resource.py?cross-origin-fetch" (expected array ["http://web-platform.test:8001/common/redirect-opt-in.py?status=307&location=https://www.web-platform.test:8445/content-security-policy/support/resource.py?cross-origin-fetch", "https://web-platform.test:9444/echo", "https://www.web-platform.test:8445/content-security-policy/support/resource.py?cross-origin-fetch", "https://www.web-platform.test:8445/content-security-policy/support/resource.py?cross-origin-xhr"] got ["http://web-platform.test:8001/common/redirect-opt-in.py?status=307&location=https://www.web-platform.test:8445/content-security-policy/support/resource.py?cross-origin-fetch", "https://www.web-platform.test:8445/content-security-policy/support/resource.py?cross-origin-fetch", "https://www.web-platform.test:8445/content-security-policy/support/resource.py?cross-origin-xhr", "wss://web-platform.test:9444/echo"])
Harness: the test ran to completion.
